requirements.txt
streamlit
groq
python-docx
pypdf
python-dotenv
requests
beautifulsoup4